backlight: Separate backlight level from exposed brightness
When we want the brightness curve to be non-linear be need two
values: those exposed to the user facing side and those values
actually set in the hardware (called `level`s).

We do the split here but use a linear mapping only. This will
change in the next commits.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1810>